saves element on touchup, properly this time, so the zindex carries over into other views via the API
